Added makeItem function to Marketplace contract with tests

+++ backend/contracts/NFT.sol
@@ -4,16 +4,16 @@ pragma solidity ^0.8.4;
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 
 contract NFT is ERC721URIStorage {
-    uint public tokenCount;
+    uint256 public tokenCount;
 
-    constructor() ERC721("DAPP NFT","DAPP"){}
+    constructor() ERC721("DAPP NFT", "DAPP") {}
 
-    function mint(string memory _tokenURI) external returns(uint){
-        tokenCount ++;
+    function mint(string memory _tokenURI) external returns (uint256) {
+        tokenCount++;
 
         _safeMint(msg.sender, tokenCount);
         _setTokenURI(tokenCount, _tokenURI);
 
-        return(tokenCount);
+        return (tokenCount);
     }
 }
